refactor(candidate): rename and restructure auth module with proper naming conventions
