+++ src/app/features/admin/interfaces/company.interface.ts
@@ -73,41 +73,6 @@ export interface ContactInfoInterface {
   github?: string;
 }
 
-export interface CandidateInterface {
-  id: string;
-
-  userId: string;
-
-  name: string;
-
-  isActive: boolean;
-
-  profileUrl?: string;
-
-  coverUrl?:string;
-
-  contactInfo?: ContactInfoInterface;
-
-  education?: string[];
-
-  skills?: string[];
-
-  experience?: string[];
-
-  resumeUrl?: string;
-
-  portfolioLinks?: string;
-
-  createdAt: Date;
-
-  updatedAt: Date;
-}
-
-export interface QueryParmsInterface {
-  page?: number
-  limit?: number
-  search?:string
-}
