+++ src/app/features/admin/services/admin-company-service.ts
@@ -2,11 +2,9 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable, } from 'rxjs';
 import {
-  CandidateInterface,
   CompanyInterface,
-  QueryParmsInterface,
 } from '../interfaces/company.interface';
-import { ApiResponce, PlainResponce } from '../../../shared/interfaces/apiresponce.interface';
+import { ApiResponce, PlainResponce, QueryParmsInterface } from '../../../shared/interfaces/apiresponce.interface';
 
 @Injectable({
   providedIn: 'root',
@@ -15,7 +13,9 @@ export class AdminCompanyService {
   private _CompnayState = new BehaviorSubject<CompanyInterface[]>([]);
   company$ = this._CompnayState.asObservable();
 
-  constructor(private readonly http: HttpClient) {}
+  constructor(
+    private readonly _http: HttpClient
+  ) {}
 
   getAllCompanies(parms:QueryParmsInterface = {}): Observable<ApiResponce<CompanyInterface[]>> {
     let httpParms = new HttpParams()
@@ -31,14 +31,14 @@ export class AdminCompanyService {
     if(parms.search){
       httpParms = httpParms.set('search',parms.search)
     }
-    return this.http.get<ApiResponce<CompanyInterface[]>>(
+    return this._http.get<ApiResponce<CompanyInterface[]>>(
       `/api/admin/getallcompany`,
       {params:httpParms, withCredentials: true }
     );
   }
 
   updateStatus(id:string):Observable<PlainResponce>{
-    return this.http.get<PlainResponce>(`/api/admin/company/updatestatus?id=${id}`,{withCredentials : true})
+    return this._http.get<PlainResponce>(`/api/admin/company/updatestatus?id=${id}`,{withCredentials : true})
   }
 
 }
